Dynamic Loading: Will now unload textures out of view
ImageGridComponent creates a range around the cursor (the current
selected game/file) and loads those files/games into memory and draws
them.  It will then unload textures of other files outside of its range
based on the direction the user is going.

If moving down the grid, ImageGrid will unload all images from 0 to the
first tile in the cursor range.
If moving up the grid, it will then unload from the end of list to the
last file in cursor range.

Hopefully this keeps vram usage under control in GridView.